requirements.txt
meshtastic
textual
pypubsub
